Decouple pydantic models and gql types
